update directory structure and utilize require
